Extract the demo view to its own file
