+++ Sources/TokamakDemo/main.swift
@@ -17,43 +17,7 @@ import TokamakDOM
 
 let document = JSObjectRef.global.document.object!
 
-struct CustomModifier: ViewModifier {
-  func body(content: Content) -> some View {
-    Text("Whole new body!")
-  }
-}
-
 let div = document.createElement!("div").object!
-let renderer = DOMRenderer(
-  ScrollView(showsIndicators: false) {
-    HStack {
-      Spacer()
-    }
-    VStack {
-      Counter(count: 5, limit: 15)
-        .padding()
-        .background(Color(red: 0.9, green: 0.9, blue: 0.9, alpha: 1.0))
-        .border(Color.red, width: 3)
-      ZStack {
-        Text("I'm on bottom")
-        Text("I'm forced to the top")
-          .zIndex(1)
-        Text("I'm on top")
-      }
-      .padding(20)
-      ForEachDemo()
-      TextDemo()
-      SVGCircle()
-        .frame(width: 25, height: 25)
-      TextFieldDemo()
-      SpacerDemo()
-      Spacer()
-      Text("Forced to bottom.")
-      EnvironmentDemo()
-        .font(.system(size: 21))
-    }
-  },
-  div
-)
+let renderer = DOMRenderer(TokamakDemoView(), div)
 
 _ = document.body.object!.appendChild!(div)
